+++ cmd/thanos/downsample.go
@@ -85,8 +85,10 @@ func RunDownsample(
 		return err
 	}
 
+	// While fetching blocks, filter out blocks that were marked for no downsample.
 	metaFetcher, err := block.NewMetaFetcher(logger, block.FetcherConcurrency, bkt, "", extprom.WrapRegistererWithPrefix("thanos_", reg), []block.MetadataFilter{
 		block.NewDeduplicateFilter(block.FetcherConcurrency),
+		downsample.NewGatherNoDownsampleMarkFilter(logger, bkt),
 	})
 	if err != nil {
 		return errors.Wrap(err, "create meta fetcher")