+++ pkg/block/metadata/markers.go
@@ -24,11 +24,15 @@ const (
 	// NoCompactMarkFilename is the known json filename for optional file storing details about why block has to be excluded from compaction.
 	// If such file is present in block dir, it means the block has to excluded from compaction (both vertical and horizontal) or rewrite (e.g deletions).
 	NoCompactMarkFilename = "no-compact-mark.json"
-
+	// NoDownsampleMarkFilename is the known json filenanme for optional file storing details about why block has to be excluded from downsampling.
+	// If such file is present in block dir, it means the block has to be excluded from downsampling.
+	NoDownsampleMarkFilename = "no-downsample-mark.json"
 	// DeletionMarkVersion1 is the version of deletion-mark file supported by Thanos.
 	DeletionMarkVersion1 = 1
 	// NoCompactMarkVersion1 is the version of no-compact-mark file supported by Thanos.
 	NoCompactMarkVersion1 = 1
+	// NoDownsampleVersion1 is the version of no-downsample-mark file supported by Thanos.
+	NoDownsampleMarkVersion1 = 1
 )
 
 var (
@@ -62,9 +66,14 @@ func (m *DeletionMark) markerFilename() string { return DeletionMarkFilename }
 // NoCompactReason is a reason for a block to be excluded from compaction.
 type NoCompactReason string
 
+// NoDownsampleReason is a reason for a block to be excluded from downsample.
+type NoDownsampleReason string
+
 const (
 	// ManualNoCompactReason is a custom reason of excluding from compaction that should be added when no-compact mark is added for unknown/user specified reason.
 	ManualNoCompactReason NoCompactReason = "manual"
+	// ManualNoDownsampleReason is a custom reason of excluding from downsample that should be added when no-downsample mark is added for unknown/user specified reason.
+	ManualNoDownsampleReason NoDownsampleReason = "manual"
 	// IndexSizeExceedingNoCompactReason is a reason of index being too big (for example exceeding 64GB limit: https://github.com/thanos-io/thanos/issues/1424)
 	// This reason can be ignored when vertical block sharding will be implemented.
 	IndexSizeExceedingNoCompactReason = "index-size-exceeding"
@@ -87,6 +96,22 @@ type NoCompactMark struct {
 }
 
 func (n *NoCompactMark) markerFilename() string { return NoCompactMarkFilename }
+
+// NoDownsampleMark marker stores reason of block being excluded from downsample if needed.
+type NoDownsampleMark struct {
+	// ID of the tsdb block.
+	ID ulid.ULID `json:"id"`
+	// Version of the file.
+	Version int `json:"version"`
+	// Details is a human readable string giving details of reason.
+	Details string `json:"details,omitempty"`
+
+	// NoDownsampleTime is a unix timestamp of when the block was marked for no downsample.
+	NoDownsampleTime int64              `json:"no_downsample_time"`
+	Reason           NoDownsampleReason `json:"reason"`
+}
+
+func (n *NoDownsampleMark) markerFilename() string { return NoDownsampleMarkFilename }
 
 // ReadMarker reads the given mark file from <dir>/<marker filename>.json in bucket.
 func ReadMarker(ctx context.Context, logger log.Logger, bkt objstore.InstrumentedBucketReader, dir string, marker Marker) error {
@@ -112,6 +137,10 @@ func ReadMarker(ctx context.Context, logger log.Logger, bkt objstore.Instrumente
 	case NoCompactMarkFilename:
 		if version := marker.(*NoCompactMark).Version; version != NoCompactMarkVersion1 {
 			return errors.Errorf("unexpected no-compact-mark file version %d, expected %d", version, NoCompactMarkVersion1)
+		}
+	case NoDownsampleMarkFilename:
+		if version := marker.(*NoDownsampleMark).Version; version != NoDownsampleMarkVersion1 {
+			return errors.Errorf("unexpected no-downsample-mark file version %d, expected %d", version, NoDownsampleMarkVersion1)
 		}
 	case DeletionMarkFilename:
 		if version := marker.(*DeletionMark).Version; version != DeletionMarkVersion1 {